Add abstraction for running a job in a Kube pod
This, in contrast to `podTemplate`, takes care of the issue where the `name`
and `label` have to be manually updated for container template updates to take
effect. It does so by automatically calculating the hash of the final arguments
passed to `podTemplate` and including that in the label.

Also, this allows getting rid of always having to include a `node` call in
`podTemplate` call and avoids the `label` and `name` value having to be
duplicated, by only expecting a `name`.