Add wrapper for configuring job's properties
This allows the trigger ("!deploy") to be configured here and not separately in
all other projects. Additionally, this allows configuring more properties as
the library keeps evolving.

This wasn't initially done, because I thought that `properties` cannot be
called in a declarative pipeline, but I tested and it works OK.

This is defined as a wrapper and not a method that returns a list that can
manually be added to an existing list of properties, because this way we can do
further modifications to the provided list of properties if needed, without
having to change Jenkinsfiles using this library. For example, if a property
we're defining conflicts with another property that may be provided by the
library user, then we can do the conflict resolution here as needed.

INF-1621