Wait for user input before deploying to beta
Currently, if a deploy is blocked behind an acceptance test run, it can take
roughly half an hour before it gets to beta. The user who started the deploy
will not likely be actively watching the build progress to see when it reaches
beta, and therefore might miss the beta deploy. If the beta deploy breaks
something, then the user might not be ready to deal with the problem or roll
back if needed.

INF-1622