Assign deployments to application-specific nodes
This fixes an issue, where this flag was used by the apply-service-settings
Jenkins job, which uses kube-envs to check if an applications manifest checksum
has changed and redeploys it if it has. The job uses this flag to deploy the
service and calculate the checksum. Because applications using Deployer didn't
use this flag, their checksums always differ from the one calculated by the
apply-service-settings job, causing the job to always redeploy them. This
caused an issue where message-hub was unexpectedly redeployed in production
during business hours.

AFAIK, the node selector was added to separate selenium pods from others in the
acceptance environment. This is not the best way to achieve that, however.
There's a way to achieve the same result without having to flag every
application deployment as is done now. The approach is described [here][1].

For now I'll avoid changing the approach and instead change Deployer to behave
as other parts of our infrastructure, by applying the node selector.

INF-1641

[1]: https://github.com/sapcc/openstack-helm#node-affinity