Specify repository name for release tool
Because we're in the git repository, we can parse the GitHub repository name
out of the origin remote URL.

This allows us to create a "deploy-pipeline-test" application in this very
repository (named "pipeline-lib"). Deploying "pipeline-lib" application to all
(including production) environments might be confusing. An explicit "test" in
the name makes it clearer - that's why the need to support application names
different from repository names.

INF-1644