Force a fresh pod for every build
The previous solution made sure that if the template was changed then the
changes were guaranteed to be picked up (which [issue][1] might be fixed by
now). But this solution (which is now also suggested in the plugin
documentation) makes sure that pods are never reused. Even for different builds
of the same job.

[1]: https://issues.jenkins-ci.org/browse/JENKINS-47759

EN-3479