Allow manual override for PR not mergeable checks
In some cases a missing review or a pending status is temporarily expected.
A review might be in progress, but the deployer might want to get the change
out to beta ASAP and then only after review continue to production. These
scenarios are now supported by not failing the job in case of problems, but
simply letting the user know that there are problems and then letting them
decide what to do.

Only send a GitHub comment that input is required when the user hasn't clicked
"Continue to beta" yet. The user is expected to be ready and attentive after
they've clicked that and so no notification is required.