Avoid mounting ~/.kube as read-only
Currently deployments are failing with `mkdir /root/.kube/schema: read-only
file system`, because we mount the secret to `~/.kube` which kubectl also tries
to use for other reasons, such as storing a schema cache (w/e that is). As
we're specifying the k8s configuration path explicitly for kubectl, we can
avoid mounting the secret to that special directory, and instead mount it to
anywhere else.

As an alternative we could only mount the specific file, instead of the entire
folder. Kubernetes supports that, but the Jenkins Kubernetes plugin currently
does not, so that option is off the table.

INF-1728