Assume deployer IAM role for deployer pods
This is required for authenticating into EKS clusters.

Potentially, the `wrapPodTemplate` method could be removed in the future.
AFAIK, nested `podTemplate` definitions already inherit all of the attributes
of the outer template, so we could perhaps simply use `inPod` or similar where
we need it, instead of requiring the users to manually wrap their templates.