Remove wrapPodTemplate
This has a couple of benefits. For one, there's one less step required for
using this library. Second, we now won't force the specific annotations etc
onto the main pod in which the job runs. So, for example, if the build needs
a different IAM role than the deployer, then now it can.

Maybe we can also get rid of `wrapProperties` at some point, but we'll see
about that.