Allow arguments after !deploy command
The arguments aren't actually used at the moment, but the same pattern can
later be used to parse the arguments. For current users this commit shouldn't
change much, except for making the pattern more lenient.

What follows are explanations of different "tricks" used for making the regex
readable.

* In Groovy "slashy strings" are multiline and avoid having to escape
  backslashes - very convenient for defining regexes.
* The lines start with the pipe operator, combined with the `stripMargin`
  function to allow the regex expression to be arbitrarily indented. This allows
  us to indent contents of groups, for example, without affecting the regex
  itself.
* `trim` is required to be able to start the regex expression from the next
  line after the variable definition. Otherwise every line's starting
  indentation would have to be so long, that the comments would very likely not
  fit into the standard max of 120 chars.

Rest is explained in the regex comments.

This was inspired by [this article][1], with improvements of using a multiline
slashly string and `stripMargin` + `trim`.

[1]: https://www.ngdc.noaa.gov/wiki/index.php/Regular_Expressions_in_Groovy