Rename no-env-lock to no-global-lock
I think this more accurately describes what the argument does. The environments
will still be locked, but just for the scope of this project, instead of
globally.

Initially considered leaving in support for `no-env-lock` for backwards
compatibility, but this argument isn't widely known or used yet, so just
renaming is probably OK at the moment.