Also reuse pre-built docker images with merge commits
If the PR is branched off from an earlier version of master, then Jenkins will
create a new merge commit for every build, to ensure latest master is included
in the build. This means that every such build will run against a commit with
a unique hash, causing `git.getShortRevision` to return a different value for
every build, even if the code is the same.

These changes check if we have a merge commit, and if we do, it uses the
revisions of master + PR head to create a unique tag for the docker image,
which doesn't change between different builds. It still changes if either
master or the PR head branch gets updated, as is expected. These combined
revision will never be used for deploys, however, for deploys we still use
a single revision, created from git hash of the head commit we get after
changing the merge commit's author to sm-deployer.

INF-1790