Add rollback link to Slack notifications
This allows for quick rollback when this deploy pipeline has finished. This is
for cases when reverting the faulty commits and deploying them is too slow. The
currently running version of the application can quickly be deployed with
a click of a button. The user will be asked for confirmation after clicking the
link.

INF-1819