Group deploy messages into threads
For every deploy, the first message sent to a channel will be sent as a simple
message as before. However, every following message will now be sent to
a thread from the original message. This minimizes the amount spam caused by
deploys.

This has two downsides. One, important messages, such as "Failure, manual
intervention required!" also get posted to the thread. I opened [a PR for the
Slack plugin][1] for this. After the PR gets merged and released, it will be
possible to broadcast some of the thread messages to the channels they're in.
Just like when you select the checkbox "Also send to #ci" when posting in
Slack.

Second, in #production, the initial messages sent to the channel includes the
words "production US", although messages in the thread will include information
about both US and EU. To avoid this, we would have to send a separate message
to #production before any of the current messages. This is not trivial,
however, and requires careful consideration before implementation.

[1]: https://github.com/jenkinsci/slack-plugin/pull/473