Notify Slack of builds being started
Builds use the 'always' strategy, when it's important to notify Slack
regardless of the outcome of the build. Often in these cases it is also useful
to get notified when the build starts. This is because these builds are likely
to have an effect before they've finished.

This change sends a message on build start and then sends the result to
a thread. If the build results in a failure, the result is also broadcasted to
the channel.

I'm not too happy with the way state is handled in this function, but don't
currently have any non-over-engineered ideas for improvement either.