Split updateStaticAssets into two methods
Sometimes only half of the current method is required. Splitting the
functionality in half to support that. Arguments and documentation were
reworded a bit for accuracy.

This could be done in a way that keeps backwards compatibility, but that's not
necessary here, because all users are known and easy to update.