Add node affinity for Jenkins worker pods
By pushing the worker pods onto separate nodes (if they're available), we avoid
them disrupting other services running on the cluster.

The worker pods can utilize the node's resources very heavily and it's very
difficult to set good resource requests and limits for them, because they're
resource usage is bursty and hard to predict. So instead of trying to give
Kubernetes enough information to schedule these pods so that they wouldn't
affect other services, we'll just separate them so that they can only affect
each other.

The affinity is set with `preferredDuringSchedulingIgnoredDuringExecution`, so
that if no nodes exist with `role=jenkins`, then the pods can still get
scheduled.

A good merge function is difficult to implement with maps and lists
intertwined, so this implementation simply overwrites the `.spec.affinity` if
it exists in the `yaml` field provided as an argument. This means that
`inPod`'s `yaml` field can still be used for most things, except for setting
affinities.