Retry merging to master
There's reason to believe that sometimes GitHub can fail the merge due to
pending status checks even though we just updated the commit statuses. The
assumption here is that it happens because the update hasn't fully propagated
yet. If that's the case, then waiting a little and trying again should help us
recover in these cases.