Remove acceptance lock usage at the end of the deploy
The acceptance environment lock is highly contended and locking it here greatly
increases the chances of congestion.

The locking is safe to avoid, because if the deploy has been tested in
production environments, it should also be safe to deploy to acceptance without
affecting other tests or services. It does require the rollout to be graceful,
but that is a basic requirement for using branch deployments in the first
place, so that should be a given.

INF-2032