Explicitly start a Slack thread in #production
Currently the deploys sometimes send two messages to #production instead of the
expected single messages that then starts a thread. This happens because US and
EU are deployed in two concurrent, `parallel` branches. Often it sending the US
messages completes before sending the EU message starts and then a single
thread is started as expected, but at other times both US and EU messages are
sent to the channel separately and then all following messages are threaded
under the US message.

In addition to avoiding confusion, this also minimizes the height of screen
space that a single deploy takes in the #production channel because only one
message is always sent instead of two sometimes. Additionally, this message is
shorter than the normal message with full environment name, previous version,
and rollback link. This further decreases the height of the message by one
line.