Acquire non-global lock for last acceptance deploy
This is a partial revert of ab8b6f4 ("Remove acceptance lock usage at the end
of the deploy", 2019-10-09). That commit removed all locking from that phase,
but this partially recovers that locking. The locking was removed, because the
global (`acceptance-environment`) lock was highly contended. With this new
implementation, that global lock is still not used and so the problem of
contention remains addressed. It recovers the "local" locking, however, because
as we've just found out, without it multiple deploys in the same project can
start deploying to acceptance at the same time, causing conflicts.