Allow running pre-deployment checks in Kubnernetes
This change will allows us to run arbitrary code in Kubernetes on the
new version of the code _before_ deploying the new version. It can be
used for tasks like validating that the new code won't make any harmful
changes is safe to deploy to the environment (think of DB migrations,
applying state to persistent storage etc).