Safely handle Slack failures
In some instances the response from Slack service is used
in further logic. Thus an empty/failed `slackSend` ends up
with a pipeline failure.
This change introduces a new wrapper function `safeSlackSend`
that will retry sending to Slack and finally ask the triggerer
whether it is OK to proceed w/o Slack message.

Part of INF-2398